Separate heroku3 monkeypatching from heroku_client
